Include the ADD command in Dockerfile to load the Process_data.ipynb
The ADD command adds the content of the local directory to the docker
image.  This sets up the docker image to explore the bat dector.

Works especially well when the labeled data set is unpacked into the
current directory.

docker build -t batdetection .

This builds a working docker image with the batdection python notebook.